refactor: useCardBrand를 유틸함수인 getCardBrand로 변경
