+++ hooks/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './validators';
 export * from './stringFormatters';
+export * from './getCardBrand';
